pyproject.toml
[project]
name = "sequoia-x"
version = "2.0.0"
requires-python = ">=3.10"
dependencies = [
    "baostock>=0.9",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "pandas>=2.0",
    "requests>=2.31",
    "akshare>=1.18.60",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "hypothesis>=6.100",
    "pytest-mock>=3.12",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
